Finished printf

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ibenli <ibenli@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 09:35:43 by ibenli            #+#    #+#             */
-/*   Updated: 2022/11/12 13:35:06 by ibenli           ###   ########.fr       */
+/*   Updated: 2022/11/12 15:34:31 by ibenli           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 # include <stdio.h>
 # include <unistd.h>
 # include <stdarg.h>
+# include <limits.h>
 
 int	ft_printf(const char *str, ...);
 
@@ -23,7 +24,7 @@ int	ft_putchar(char c);
 int	ft_putnbr(int n);
 int	ft_putstr(char *str);
 int	ft_putnbr_unsigned(unsigned int n);
-int ft_puthexa(char c, unsigned int n);
-int ft_putadress(char c, unsigned int n);
+int ft_puthexa(char c, unsigned long n);
+int ft_putadress(char c, unsigned long n);
 
 #endif
